requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66

rank-bm25>=0.2.2
faiss-cpu>=1.8.0

sentence-transformers>=3.0.0
transformers>=4.43
torch>=2.2

rich>=13.7
pyyaml>=6.0.1
accelerate

streamlit